transaksi

+++ database/migrations/2026_02_08_154725_update_kt_transaksi_table.php
@@ -12,12 +12,15 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('kt_transaksi', function (Blueprint $table) {
-            // hapus FK & kolom id_users
-            $table->dropForeign(['id_users']);
-            $table->dropColumn('id_users');
-
-            // tambah customer text
+            if (Schema::hasColumn('kt_transaksi', 'id_users')) {
+                $table->dropForeign(['id_users']);
+                $table->dropColumn('id_users');
+            }
             $table->string('customer')->nullable()->after('id_supplier');
+            $table->date('jatuh_tempo')->nullable()->after('tanggal');
+            $table->enum('status_pembayaran', ['BELUM_LUNAS', 'LUNAS'])
+                ->default('BELUM_LUNAS')
+                ->after('total_bayar');
         });
     }
 
@@ -27,11 +30,14 @@ return new class extends Migration
     public function down(): void
     {
         Schema::table('kt_transaksi', function (Blueprint $table) {
-            // hapus customer
-            $table->dropColumn('customer');
+            $table->dropColumn([
+                'customer',
+                'jatuh_tempo',
+                'status_pembayaran'
+            ]);
 
             // balikin id_users
-            $table->unsignedBigInteger('id_users');
+            $table->unsignedBigInteger('id_users')->nullable();
 
             $table->foreign('id_users')
                 ->references('id')
